REST 수업

+++ study/src/main/java/com/study/project/controller/BoardController.java
@@ -6,18 +6,17 @@ import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
 
 import com.study.project.service.BoardServiceInter;
 
 @Controller
 public class BoardController {
+	
 	
 	@Autowired
 	public BoardServiceInter service;
@@ -25,6 +24,7 @@ public class BoardController {
 	@RequestMapping("list")
 	@ResponseBody
 	public List<Map<String, Object>> list() {
+		System.out.println("===NOT REST LIST===");
 		
 		List<Map<String, Object>> list = service.list();
 		
@@ -34,7 +34,7 @@ public class BoardController {
 	@RequestMapping("/detail/{num}")
 	@ResponseBody
 	public Map<String, Object> detail(@PathVariable(name="num") int num) {
-		
+		System.out.println("===NOT REST DETAIL===");
 		Map<String, Object> board = service.detail(num);
 		
 		return board;
@@ -43,6 +43,7 @@ public class BoardController {
 	@RequestMapping("insert")
 	@ResponseBody
 	public Map<String, Object> insert(@RequestBody Map<String, Object> insertMap) {
+		System.out.println("===NOT REST INSERT===");
 		int insert = service.insert(insertMap);
 		Map<String, Object> status = new HashMap<String, Object>();
 		
@@ -59,6 +60,7 @@ public class BoardController {
 	@RequestMapping("update")
 	@ResponseBody
 	public Map<String, Object> update(@RequestBody Map<String, Object> updateMap) {
+		System.out.println("===NOT REST UPDATE===");
 		int update = service.update(updateMap);
 		Map<String, Object> status = new HashMap<String, Object>();
 		
@@ -74,9 +76,18 @@ public class BoardController {
 	
 	@RequestMapping("delete")
 	@ResponseBody
-	public Map<String, Object> delete(@RequestBody Map<String, List<Integer>> deleteList) {
-		int delete = service.delete(deleteList);
+	public Map<String, Object> delete(@RequestBody Map<String, List<Integer>> deleteMap) {
+		System.out.println("===NOT REST DELETE===");
 		Map<String, Object> status = new HashMap<>();
+		
+		List<Integer> deleteList = deleteMap.get("boardNums");
+		
+		if(CollectionUtils.isEmpty(deleteList)) {
+			status.put("stus", "fail");
+			return status;
+		}
+		
+		int delete = service.delete(deleteList);
 		
 		if (delete == 0) {
 			status.put("stus", "fail");
